+++ CHANGELOG.md
@@ -3,7 +3,7 @@ version 3.441
 check ip for admin login
 fix link orders dashboard
 Update ckeditor
-
+Stripe
 
 --------------------
 version 3.440